backend/requirements.txt
python-dotenv

fastapi[standard]
python-multipart
uvicorn

supabase

pydantic-ai

replicate
